+++ src/lib/server/email/password-reset.ts
@@ -1,5 +1,5 @@
-import { DOMAIN, SENDGRID_SENDER } from '$env/static/private'
-import { sendMessage } from '$lib/server/sendgrid'
+import { DOMAIN, EMAIL } from '$env/static/private'
+import { sendMessage } from '$lib/server/brevo'
 
 /**
  * Sends a password reset email containing a link with a one-time reset token.
@@ -9,11 +9,11 @@ import { sendMessage } from '$lib/server/sendgrid'
  */
 export const sendPasswordResetEmail = async (toEmail: string, token: string) => {
 	await sendMessage({
-		to: { email: toEmail },
-		from: SENDGRID_SENDER,
+		to: [{ email: toEmail }],
+		sender: { email: EMAIL },
 		subject: 'Password reset',
-		categories: ['account'],
-		html: `
+		tags: ['account'],
+		htmlContent: `
       <p><a href="${DOMAIN}/auth/reset/${token}">Reset my password</a>. Your browser will open and ask you to
       provide a new password with a confirmation then redirect you to your login page.</p>
     `